Lint cleanup: isort/black/ruff autofix + targeted manual fixes
Mechanical cleanup:
- ruff --fix removed ~575 unused imports across MoE-v2 / kernels / train.
- isort reorganized import blocks.
- black reformatted 71 files.

Targeted manual fixes (uncovered during cleanup):
- nn/moe/v2/block.py: moved process-local cache locks below all imports
  (resolves 28 E402 warnings in the file).
- nn/moe/v2/ep_no_sync_rowwise_helpers.py: added missing TYPE_CHECKING
  import of ReduceType (used as string-quoted annotation).
- train/checkpoint.py: scatter_object → broadcast_object at lines 591
  and 656. scatter_object was renamed in main PR #471 and never updated
  on tianhua's branch - would have crashed on Trainer.eval_checkpoints
  when called.
- train_module/transformer/moe_train_module.py: removed duplicate
  DataParallelType import (was imported from both .parallel and
  .parallel.data_parallel).
- nn/functional/cross_entropy_loss.py: noqa-marked the
  cce_compute_z_loss import (used by callers via re-export).
- optim/zero_optim.py: moved the noqa-F841 comment back onto the line
  black had moved off of.

Known pre-existing tianhua-side issues NOT fixed (would need real code
changes, not lint passes):
- nn/attention/__init__.py: MultiheadLatentAttention uses repeat_kv,
  dispatch_flash_attn, dispatch_ring_flash_attn, cu_doc_lens without
  imports/definitions. MLA isn't exercised by any current training
  script, so the bugs are dormant.
- nn/moe/v2/model.py: apply_ep is defined twice (stub at line 495 +
  real impl at line 546). Python takes the second one but the stub is
  dead code - ambiguous what was intended.
- Various F841 dbg_mem_* unused locals in moe_optimizer/moe_optimizer_old
  - debug instrumentation with the print statements commented out.

Ruff count: 1044 → 443 errors. The 443 remaining are mostly E402
(intentional import patterns in training scripts where setup code runs
before imports, ~392 of those) plus the known issues above.

+++ src/test/nn/moe/mxfp8_tensor_subclass_test.py
@@ -59,7 +59,9 @@ def test_olmo_mxfp8_tensor_can_be_backward_gradient_object():
 
         @staticmethod
         def backward(ctx, grad_out):  # type: ignore[override]
-            hp_grad = torch.ones(ctx.shape, dtype=ctx.dtype, device=ctx.device) * grad_out.to(ctx.dtype)
+            hp_grad = torch.ones(ctx.shape, dtype=ctx.dtype, device=ctx.device) * grad_out.to(
+                ctx.dtype
+            )
             return OlmoMXFP8Tensor.from_hp(hp_grad, prefer_triton=False)
 
     x = torch.randn(4, 64, dtype=torch.bfloat16, requires_grad=True)
